+++ src/main/java/roomescape/user/dto/TimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.user.dto;
 
 import java.time.LocalTime;
